Reimplemented three separate components
